# lib/api-zod/src/generated/types/countryDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Global Scholar & Migration Guide API
 * OpenAPI spec version: 0.1.0
 */
import type { Opportunity } from "./opportunity";

export interface CountryDetail {
  code: string;
  name: string;
  nameAr: string;
  flag: string;
  region: string;
  description: string;
  opportunities: Opportunity[];
  opportunityCount: number;
  scholarshipCount: number;
  migrationCount: number;
}