# lib/api-zod/src/generated/types/createApplicationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Global Scholar & Migration Guide API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateApplicationBody {
  userId: string;
  opportunityId: number;
  status?: string;
  notes?: string;
}