# lib/api-zod/src/generated/types/opportunityList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Global Scholar & Migration Guide API
 * OpenAPI spec version: 0.1.0
 */
import type { Opportunity } from "./opportunity";

export interface OpportunityList {
  items: Opportunity[];
  total: number;
  page: number;
  pageSize: number;
}