# lib/api-zod/src/generated/types/updateApplicationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Global Scholar & Migration Guide API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateApplicationBody {
  status?: string;
  notes?: string;
}